requirements.txt
aiogram==3.0.0b7
emoji==2.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.5.post1
aiosqlite==0.18.0